LPS-126971 Move getProps* test methods to new class ContentDashboardAdminPortletGetPropsTest
