+++ modules/apps/content-dashboard/content-dashboard-test/src/testIntegration/java/com/liferay/content/dashboard/web/internal/portlet/test/ContentDashboardAdminPortletTest.java
@@ -31,7 +31,6 @@ import com.liferay.journal.service.JournalArticleLocalService;
 import com.liferay.journal.test.util.JournalTestUtil;
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.dao.search.SearchContainer;
-import com.liferay.portal.kernel.json.JSONFactoryUtil;
 import com.liferay.portal.kernel.json.JSONUtil;
 import com.liferay.portal.kernel.language.constants.LanguageConstants;
 import com.liferay.portal.kernel.model.Company;
@@ -275,1002 +274,6 @@ public class ContentDashboardAdminPortletTest {
 				HtmlUtil.escapeJS(
 					"mvcRenderCommandName=/content_dashboard" +
 						"/edit_content_dashboard_configuration")));
-	}
-
-	@Test
-	public void testGetProps() throws Exception {
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(
-				_company.getCompanyId(), _company.getGroupId(),
-				_user.getUserId());
-
-		AssetVocabulary assetVocabulary =
-			_assetVocabularyLocalService.fetchGroupVocabulary(
-				serviceContext.getScopeGroupId(), "audience");
-
-		AssetCategory assetCategory = _assetCategoryLocalService.addCategory(
-			_user.getUserId(), _company.getGroupId(),
-			RandomTestUtil.randomString(), assetVocabulary.getVocabularyId(),
-			serviceContext);
-
-		AssetVocabulary childAssetVocabulary =
-			_assetVocabularyLocalService.fetchGroupVocabulary(
-				serviceContext.getScopeGroupId(), "stage");
-
-		AssetCategory childAssetCategory =
-			_assetCategoryLocalService.addCategory(
-				_user.getUserId(), _company.getGroupId(),
-				RandomTestUtil.randomString(),
-				childAssetVocabulary.getVocabularyId(), serviceContext);
-
-		try {
-			JournalTestUtil.addArticle(
-				_group.getGroupId(), 0,
-				ServiceContextTestUtil.getServiceContext(
-					_group.getGroupId(), _user.getUserId(),
-					new long[] {
-						assetCategory.getCategoryId(),
-						childAssetCategory.getCategoryId()
-					}));
-
-			Map<String, Object> data = _getData(
-				_getMockLiferayPortletRenderRequest());
-
-			Map<String, Object> props = (Map<String, Object>)data.get("props");
-
-			Assert.assertEquals(
-				JSONUtil.put(
-					JSONUtil.put(
-						"categories",
-						JSONUtil.put(
-							JSONUtil.put(
-								"key",
-								String.valueOf(
-									childAssetCategory.getCategoryId())
-							).put(
-								"name",
-								childAssetCategory.getTitle(LocaleUtil.US)
-							).put(
-								"value", 1L
-							).put(
-								"vocabularyName",
-								childAssetVocabulary.getTitle(LocaleUtil.US)
-							))
-					).put(
-						"key", String.valueOf(assetCategory.getCategoryId())
-					).put(
-						"name", assetCategory.getTitle(LocaleUtil.US)
-					).put(
-						"value", 1L
-					).put(
-						"vocabularyName",
-						assetVocabulary.getTitle(LocaleUtil.US)
-					)
-				).toString(),
-				String.valueOf(props.get("vocabularies")));
-		}
-		finally {
-			_assetCategoryLocalService.deleteAssetCategory(assetCategory);
-			_assetCategoryLocalService.deleteAssetCategory(childAssetCategory);
-		}
-	}
-
-	@Test
-	public void testGetPropsWithAssetCategoriesSortedByKey() throws Exception {
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(
-				_company.getGroupId(), _user.getUserId());
-
-		AssetVocabulary assetVocabulary =
-			_assetVocabularyLocalService.addVocabulary(
-				_user.getUserId(), _company.getGroupId(), "vocabulary",
-				serviceContext);
-
-		AssetCategory assetCategory1 = _assetCategoryLocalService.addCategory(
-			_user.getUserId(), _company.getGroupId(), "category-1",
-			assetVocabulary.getVocabularyId(), serviceContext);
-		AssetCategory assetCategory2 = _assetCategoryLocalService.addCategory(
-			_user.getUserId(), _company.getGroupId(), "category-2",
-			assetVocabulary.getVocabularyId(), serviceContext);
-
-		try {
-			JournalTestUtil.addArticle(
-				_group.getGroupId(), 0,
-				ServiceContextTestUtil.getServiceContext(
-					_group.getGroupId(), _user.getUserId(),
-					new long[] {
-						assetCategory1.getCategoryId(),
-						assetCategory2.getCategoryId()
-					}));
-			JournalTestUtil.addArticle(
-				_group.getGroupId(), 0,
-				ServiceContextTestUtil.getServiceContext(
-					_group.getGroupId(), _user.getUserId(),
-					new long[] {assetCategory2.getCategoryId()}));
-
-			Map<String, Object> data = _getData(
-				_getMockLiferayPortletRenderRequest(
-					new String[] {assetVocabulary.getName()},
-					LocaleUtil.getSiteDefault()));
-
-			Map<String, Object> props = (Map<String, Object>)data.get("props");
-
-			Assert.assertEquals(
-				JSONUtil.putAll(
-					JSONUtil.put(
-						"key", String.valueOf(assetCategory1.getCategoryId())
-					).put(
-						"name", "category-1"
-					).put(
-						"value", 1L
-					).put(
-						"vocabularyName", "vocabulary"
-					),
-					JSONUtil.put(
-						"key", String.valueOf(assetCategory2.getCategoryId())
-					).put(
-						"name", "category-2"
-					).put(
-						"value", 2L
-					).put(
-						"vocabularyName", "vocabulary"
-					)
-				).toString(),
-				String.valueOf(props.get("vocabularies")));
-		}
-		finally {
-			_assetVocabularyLocalService.deleteAssetVocabulary(assetVocabulary);
-		}
-	}
-
-	@Test
-	public void testGetPropsWithChildAssetCategoriesSortedByKey()
-		throws Exception {
-
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(
-				_company.getGroupId(), _user.getUserId());
-
-		AssetVocabulary assetVocabulary =
-			_assetVocabularyLocalService.addVocabulary(
-				_user.getUserId(), _company.getGroupId(), "vocabulary",
-				serviceContext);
-
-		AssetCategory assetCategory = _assetCategoryLocalService.addCategory(
-			_user.getUserId(), _company.getGroupId(), "category",
-			assetVocabulary.getVocabularyId(), serviceContext);
-
-		AssetVocabulary childAssetVocabulary =
-			_assetVocabularyLocalService.addVocabulary(
-				_user.getUserId(), _company.getGroupId(), "child-vocabulary",
-				serviceContext);
-
-		AssetCategory childAssetCategory1 =
-			_assetCategoryLocalService.addCategory(
-				_user.getUserId(), _company.getGroupId(), "child-category-1",
-				childAssetVocabulary.getVocabularyId(), serviceContext);
-		AssetCategory childAssetCategory2 =
-			_assetCategoryLocalService.addCategory(
-				_user.getUserId(), _company.getGroupId(), "child-category-2",
-				childAssetVocabulary.getVocabularyId(), serviceContext);
-
-		try {
-			JournalTestUtil.addArticle(
-				_group.getGroupId(), 0,
-				ServiceContextTestUtil.getServiceContext(
-					_group.getGroupId(), _user.getUserId(),
-					new long[] {
-						assetCategory.getCategoryId(),
-						childAssetCategory1.getCategoryId(),
-						childAssetCategory2.getCategoryId()
-					}));
-			JournalTestUtil.addArticle(
-				_group.getGroupId(), 0,
-				ServiceContextTestUtil.getServiceContext(
-					_group.getGroupId(), _user.getUserId(),
-					new long[] {
-						assetCategory.getCategoryId(),
-						childAssetCategory2.getCategoryId()
-					}));
-
-			Map<String, Object> data = _getData(
-				_getMockLiferayPortletRenderRequest(
-					new String[] {
-						assetVocabulary.getName(),
-						childAssetVocabulary.getName()
-					},
-					LocaleUtil.getSiteDefault()));
-
-			Map<String, Object> props = (Map<String, Object>)data.get("props");
-
-			Assert.assertEquals(
-				JSONUtil.putAll(
-					JSONUtil.put(
-						"categories",
-						JSONUtil.putAll(
-							JSONUtil.put(
-								"key",
-								String.valueOf(
-									childAssetCategory1.getCategoryId())
-							).put(
-								"name", "child-category-1"
-							).put(
-								"value", 1L
-							).put(
-								"vocabularyName", "child-vocabulary"
-							),
-							JSONUtil.put(
-								"key",
-								String.valueOf(
-									childAssetCategory2.getCategoryId())
-							).put(
-								"name", "child-category-2"
-							).put(
-								"value", 2L
-							).put(
-								"vocabularyName", "child-vocabulary"
-							))
-					).put(
-						"key", String.valueOf(assetCategory.getCategoryId())
-					).put(
-						"name", "category"
-					).put(
-						"value", 2L
-					).put(
-						"vocabularyName", "vocabulary"
-					)
-				).toString(),
-				String.valueOf(props.get("vocabularies")));
-		}
-		finally {
-			_assetVocabularyLocalService.deleteVocabulary(assetVocabulary);
-			_assetVocabularyLocalService.deleteVocabulary(childAssetVocabulary);
-		}
-	}
-
-	@Test
-	public void testGetPropsWithChildNoneAssetCategory() throws Exception {
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(
-				_company.getCompanyId(), _company.getGroupId(),
-				_user.getUserId());
-
-		AssetVocabulary assetVocabulary =
-			_assetVocabularyLocalService.fetchGroupVocabulary(
-				serviceContext.getScopeGroupId(), "audience");
-
-		AssetCategory assetCategory = _assetCategoryLocalService.addCategory(
-			_user.getUserId(), _company.getGroupId(),
-			RandomTestUtil.randomString(), assetVocabulary.getVocabularyId(),
-			serviceContext);
-
-		AssetVocabulary childAssetVocabulary =
-			_assetVocabularyLocalService.fetchGroupVocabulary(
-				serviceContext.getScopeGroupId(), "stage");
-
-		AssetCategory childAssetCategory =
-			_assetCategoryLocalService.addCategory(
-				_user.getUserId(), _company.getGroupId(),
-				RandomTestUtil.randomString(),
-				childAssetVocabulary.getVocabularyId(), serviceContext);
-
-		try {
-			JournalTestUtil.addArticle(
-				_group.getGroupId(), 0,
-				ServiceContextTestUtil.getServiceContext(
-					_group.getGroupId(), _user.getUserId(),
-					new long[] {assetCategory.getCategoryId()}));
-			JournalTestUtil.addArticle(
-				_group.getGroupId(), 0,
-				ServiceContextTestUtil.getServiceContext(
-					_group.getGroupId(), _user.getUserId(),
-					new long[] {
-						assetCategory.getCategoryId(),
-						childAssetCategory.getCategoryId()
-					}));
-
-			Map<String, Object> data = _getData(
-				_getMockLiferayPortletRenderRequest());
-
-			Map<String, Object> props = (Map<String, Object>)data.get("props");
-
-			Assert.assertEquals(
-				JSONUtil.putAll(
-					JSONUtil.put(
-						"categories",
-						JSONUtil.putAll(
-							JSONUtil.put(
-								"key",
-								String.valueOf(
-									childAssetCategory.getCategoryId())
-							).put(
-								"name",
-								childAssetCategory.getTitle(LocaleUtil.US)
-							).put(
-								"value", 1L
-							).put(
-								"vocabularyName",
-								childAssetVocabulary.getTitle(LocaleUtil.US)
-							),
-							JSONUtil.put(
-								"key", "none"
-							).put(
-								"name", "No Stage Specified"
-							).put(
-								"value", 1L
-							).put(
-								"vocabularyName",
-								childAssetVocabulary.getTitle(LocaleUtil.US)
-							))
-					).put(
-						"key", String.valueOf(assetCategory.getCategoryId())
-					).put(
-						"name", assetCategory.getTitle(LocaleUtil.US)
-					).put(
-						"value", 2L
-					).put(
-						"vocabularyName",
-						assetVocabulary.getTitle(LocaleUtil.US)
-					)
-				).toString(),
-				String.valueOf(props.get("vocabularies")));
-		}
-		finally {
-			_assetCategoryLocalService.deleteAssetCategory(assetCategory);
-			_assetCategoryLocalService.deleteAssetCategory(childAssetCategory);
-		}
-	}
-
-	@Test
-	public void testGetPropsWithChildNoneAssetCategoryAndNoneAssetCategory()
-		throws Exception {
-
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(
-				_company.getCompanyId(), _company.getGroupId(),
-				_user.getUserId());
-
-		AssetVocabulary assetVocabulary =
-			_assetVocabularyLocalService.fetchGroupVocabulary(
-				serviceContext.getScopeGroupId(), "audience");
-
-		AssetCategory assetCategory1 = _assetCategoryLocalService.addCategory(
-			_user.getUserId(), _company.getGroupId(), "A1",
-			assetVocabulary.getVocabularyId(), serviceContext);
-		AssetCategory assetCategory2 = _assetCategoryLocalService.addCategory(
-			_user.getUserId(), _company.getGroupId(), "A2",
-			assetVocabulary.getVocabularyId(), serviceContext);
-
-		AssetVocabulary childAssetVocabulary =
-			_assetVocabularyLocalService.fetchGroupVocabulary(
-				serviceContext.getScopeGroupId(), "stage");
-
-		AssetCategory childAssetCategory =
-			_assetCategoryLocalService.addCategory(
-				_user.getUserId(), _company.getGroupId(), "S1",
-				childAssetVocabulary.getVocabularyId(), serviceContext);
-
-		try {
-			JournalTestUtil.addArticle(
-				_group.getGroupId(), 0,
-				ServiceContextTestUtil.getServiceContext(
-					_group.getGroupId(), _user.getUserId(),
-					new long[] {
-						assetCategory1.getCategoryId(),
-						childAssetCategory.getCategoryId()
-					}));
-			JournalTestUtil.addArticle(
-				_group.getGroupId(), 0,
-				ServiceContextTestUtil.getServiceContext(
-					_group.getGroupId(), _user.getUserId(),
-					new long[] {
-						assetCategory1.getCategoryId(),
-						assetCategory2.getCategoryId()
-					}));
-			JournalTestUtil.addArticle(
-				_group.getGroupId(), 0,
-				ServiceContextTestUtil.getServiceContext(
-					_group.getGroupId(), _user.getUserId(),
-					new long[] {childAssetCategory.getCategoryId()}));
-
-			Map<String, Object> data = _getData(
-				_getMockLiferayPortletRenderRequest());
-
-			Map<String, Object> props = (Map<String, Object>)data.get("props");
-
-			Assert.assertEquals(
-				JSONUtil.putAll(
-					JSONUtil.put(
-						"categories",
-						JSONUtil.putAll(
-							JSONUtil.put(
-								"key",
-								String.valueOf(
-									childAssetCategory.getCategoryId())
-							).put(
-								"name", "S1"
-							).put(
-								"value", 1L
-							).put(
-								"vocabularyName",
-								childAssetVocabulary.getTitle(LocaleUtil.US)
-							),
-							JSONUtil.put(
-								"key", "none"
-							).put(
-								"name", "No Stage Specified"
-							).put(
-								"value", 1L
-							).put(
-								"vocabularyName",
-								childAssetVocabulary.getTitle(LocaleUtil.US)
-							))
-					).put(
-						"key", String.valueOf(assetCategory1.getCategoryId())
-					).put(
-						"name", "A1"
-					).put(
-						"value", 2L
-					).put(
-						"vocabularyName",
-						assetVocabulary.getTitle(LocaleUtil.US)
-					),
-					JSONUtil.put(
-						"categories",
-						JSONUtil.putAll(
-							JSONUtil.put(
-								"key",
-								String.valueOf(
-									childAssetCategory.getCategoryId())
-							).put(
-								"name", "S1"
-							).put(
-								"value", 0L
-							).put(
-								"vocabularyName",
-								childAssetVocabulary.getTitle(LocaleUtil.US)
-							),
-							JSONUtil.put(
-								"key", "none"
-							).put(
-								"name", "No Stage Specified"
-							).put(
-								"value", 1L
-							).put(
-								"vocabularyName",
-								childAssetVocabulary.getTitle(LocaleUtil.US)
-							))
-					).put(
-						"key", String.valueOf(assetCategory2.getCategoryId())
-					).put(
-						"name", "A2"
-					).put(
-						"value", 1L
-					).put(
-						"vocabularyName",
-						assetVocabulary.getTitle(LocaleUtil.US)
-					),
-					JSONUtil.put(
-						"categories",
-						JSONUtil.put(
-							JSONUtil.put(
-								"key",
-								String.valueOf(
-									childAssetCategory.getCategoryId())
-							).put(
-								"name", "S1"
-							).put(
-								"value", 1L
-							).put(
-								"vocabularyName",
-								childAssetVocabulary.getTitle(LocaleUtil.US)
-							))
-					).put(
-						"key", "none"
-					).put(
-						"name", "No Audience Specified"
-					).put(
-						"value", 1L
-					).put(
-						"vocabularyName",
-						assetVocabulary.getTitle(LocaleUtil.US)
-					)
-				).toString(),
-				String.valueOf(props.get("vocabularies")));
-		}
-		finally {
-			_assetCategoryLocalService.deleteAssetCategory(assetCategory1);
-			_assetCategoryLocalService.deleteAssetCategory(childAssetCategory);
-		}
-	}
-
-	@Test
-	public void testGetPropsWithEmptyChildAssetCategoriesSortedByKey()
-		throws Exception {
-
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(
-				_company.getGroupId(), _user.getUserId());
-
-		AssetVocabulary assetVocabulary =
-			_assetVocabularyLocalService.addVocabulary(
-				_user.getUserId(), _company.getGroupId(), "vocabulary",
-				serviceContext);
-
-		AssetCategory assetCategory1 = _assetCategoryLocalService.addCategory(
-			_user.getUserId(), _company.getGroupId(), "category-1",
-			assetVocabulary.getVocabularyId(), serviceContext);
-		AssetCategory assetCategory2 = _assetCategoryLocalService.addCategory(
-			_user.getUserId(), _company.getGroupId(), "category-2",
-			assetVocabulary.getVocabularyId(), serviceContext);
-
-		AssetVocabulary childAssetVocabulary =
-			_assetVocabularyLocalService.addVocabulary(
-				_user.getUserId(), _company.getGroupId(), "child-vocabulary",
-				serviceContext);
-
-		AssetCategory childAssetCategory1 =
-			_assetCategoryLocalService.addCategory(
-				_user.getUserId(), _company.getGroupId(), "child-category-1",
-				childAssetVocabulary.getVocabularyId(), serviceContext);
-		AssetCategory childAssetCategory2 =
-			_assetCategoryLocalService.addCategory(
-				_user.getUserId(), _company.getGroupId(), "child-category-2",
-				childAssetVocabulary.getVocabularyId(), serviceContext);
-
-		try {
-			JournalTestUtil.addArticle(
-				_group.getGroupId(), 0,
-				ServiceContextTestUtil.getServiceContext(
-					_group.getGroupId(), _user.getUserId(),
-					new long[] {
-						assetCategory1.getCategoryId(),
-						childAssetCategory2.getCategoryId()
-					}));
-			JournalTestUtil.addArticle(
-				_group.getGroupId(), 0,
-				ServiceContextTestUtil.getServiceContext(
-					_group.getGroupId(), _user.getUserId(),
-					new long[] {
-						assetCategory2.getCategoryId(),
-						childAssetCategory1.getCategoryId()
-					}));
-
-			Map<String, Object> data = _getData(
-				_getMockLiferayPortletRenderRequest(
-					new String[] {
-						assetVocabulary.getName(),
-						childAssetVocabulary.getName()
-					},
-					LocaleUtil.getSiteDefault()));
-
-			Map<String, Object> props = (Map<String, Object>)data.get("props");
-
-			Assert.assertEquals(
-				JSONUtil.putAll(
-					JSONUtil.put(
-						"categories",
-						JSONUtil.putAll(
-							JSONUtil.put(
-								"key",
-								String.valueOf(
-									childAssetCategory1.getCategoryId())
-							).put(
-								"name", "child-category-1"
-							).put(
-								"value", 0L
-							).put(
-								"vocabularyName", "child-vocabulary"
-							),
-							JSONUtil.put(
-								"key",
-								String.valueOf(
-									childAssetCategory2.getCategoryId())
-							).put(
-								"name", "child-category-2"
-							).put(
-								"value", 1L
-							).put(
-								"vocabularyName", "child-vocabulary"
-							))
-					).put(
-						"key", String.valueOf(assetCategory1.getCategoryId())
-					).put(
-						"name", "category-1"
-					).put(
-						"value", 1L
-					).put(
-						"vocabularyName", "vocabulary"
-					),
-					JSONUtil.put(
-						"categories",
-						JSONUtil.putAll(
-							JSONUtil.put(
-								"key",
-								String.valueOf(
-									childAssetCategory1.getCategoryId())
-							).put(
-								"name", "child-category-1"
-							).put(
-								"value", 1L
-							).put(
-								"vocabularyName", "child-vocabulary"
-							),
-							JSONUtil.put(
-								"key",
-								String.valueOf(
-									childAssetCategory2.getCategoryId())
-							).put(
-								"name", "child-category-2"
-							).put(
-								"value", 0L
-							).put(
-								"vocabularyName", "child-vocabulary"
-							))
-					).put(
-						"key", String.valueOf(assetCategory2.getCategoryId())
-					).put(
-						"name", "category-2"
-					).put(
-						"value", 1L
-					).put(
-						"vocabularyName", "vocabulary"
-					)
-				).toString(),
-				String.valueOf(props.get("vocabularies")));
-		}
-		finally {
-			_assetVocabularyLocalService.deleteVocabulary(assetVocabulary);
-			_assetVocabularyLocalService.deleteVocabulary(childAssetVocabulary);
-		}
-	}
-
-	@Test
-	public void testGetPropsWithMissingCategorizedJournalArticle()
-		throws Exception {
-
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(
-				_company.getCompanyId(), _company.getGroupId(),
-				_user.getUserId());
-
-		AssetVocabulary assetVocabulary =
-			_assetVocabularyLocalService.fetchGroupVocabulary(
-				serviceContext.getScopeGroupId(), "audience");
-
-		AssetCategory assetCategory = _assetCategoryLocalService.addCategory(
-			_user.getUserId(), _company.getGroupId(),
-			RandomTestUtil.randomString(), assetVocabulary.getVocabularyId(),
-			serviceContext);
-
-		AssetVocabulary childAssetVocabulary =
-			_assetVocabularyLocalService.fetchGroupVocabulary(
-				serviceContext.getScopeGroupId(), "stage");
-
-		AssetCategory childAssetCategory =
-			_assetCategoryLocalService.addCategory(
-				_user.getUserId(), _company.getGroupId(),
-				RandomTestUtil.randomString(),
-				childAssetVocabulary.getVocabularyId(), serviceContext);
-
-		try {
-			JournalTestUtil.addArticle(
-				_user.getUserId(), _group.getGroupId(), 0);
-
-			Map<String, Object> data = _getData(
-				_getMockLiferayPortletRenderRequest());
-
-			Map<String, Object> props = (Map<String, Object>)data.get("props");
-
-			Assert.assertEquals(
-				String.valueOf(JSONFactoryUtil.createJSONArray()),
-				String.valueOf(props.get("vocabularies")));
-		}
-		finally {
-			_assetCategoryLocalService.deleteAssetCategory(assetCategory);
-			_assetCategoryLocalService.deleteAssetCategory(childAssetCategory);
-		}
-	}
-
-	@Test
-	public void testGetPropsWithMissingCategorizedJournalArticleWithAssetCategory()
-		throws Exception {
-
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(
-				_company.getCompanyId(), _company.getGroupId(),
-				_user.getUserId());
-
-		AssetVocabulary assetVocabulary =
-			_assetVocabularyLocalService.fetchGroupVocabulary(
-				serviceContext.getScopeGroupId(), "audience");
-
-		AssetCategory assetCategory = _assetCategoryLocalService.addCategory(
-			_user.getUserId(), _company.getGroupId(),
-			RandomTestUtil.randomString(), assetVocabulary.getVocabularyId(),
-			serviceContext);
-
-		AssetVocabulary childAssetVocabulary =
-			_assetVocabularyLocalService.fetchGroupVocabulary(
-				serviceContext.getScopeGroupId(), "stage");
-
-		AssetCategory childAssetCategory =
-			_assetCategoryLocalService.addCategory(
-				_user.getUserId(), _company.getGroupId(),
-				RandomTestUtil.randomString(),
-				childAssetVocabulary.getVocabularyId(), serviceContext);
-
-		try {
-			JournalTestUtil.addArticle(
-				_group.getGroupId(), 0,
-				ServiceContextTestUtil.getServiceContext(
-					_group.getGroupId(), _user.getUserId(),
-					new long[] {childAssetCategory.getCategoryId()}));
-
-			Map<String, Object> data = _getData(
-				_getMockLiferayPortletRenderRequest());
-
-			Map<String, Object> props = (Map<String, Object>)data.get("props");
-
-			Assert.assertEquals(
-				JSONUtil.put(
-					JSONUtil.put(
-						"key",
-						String.valueOf(childAssetCategory.getCategoryId())
-					).put(
-						"name", childAssetCategory.getTitle(LocaleUtil.US)
-					).put(
-						"value", 1L
-					).put(
-						"vocabularyName",
-						childAssetVocabulary.getTitle(LocaleUtil.US)
-					)
-				).toString(),
-				String.valueOf(props.get("vocabularies")));
-		}
-		finally {
-			_assetCategoryLocalService.deleteAssetCategory(assetCategory);
-			_assetCategoryLocalService.deleteAssetCategory(childAssetCategory);
-		}
-	}
-
-	@Test
-	public void testGetPropsWithMissingCategorizedJournalArticleWithChildAssetCategory()
-		throws Exception {
-
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(
-				_company.getGroupId(), _user.getUserId());
-
-		AssetVocabulary assetVocabulary =
-			_assetVocabularyLocalService.addVocabulary(
-				_user.getUserId(), _company.getGroupId(), "vocabulary",
-				serviceContext);
-
-		AssetCategory assetCategory = _assetCategoryLocalService.addCategory(
-			_user.getUserId(), _company.getGroupId(), "category",
-			assetVocabulary.getVocabularyId(), serviceContext);
-
-		AssetVocabulary childAssetVocabulary =
-			_assetVocabularyLocalService.addVocabulary(
-				_user.getUserId(), _company.getGroupId(),
-				RandomTestUtil.randomString(), serviceContext);
-
-		_assetCategoryLocalService.addCategory(
-			_user.getUserId(), _company.getGroupId(),
-			RandomTestUtil.randomString(),
-			childAssetVocabulary.getVocabularyId(), serviceContext);
-
-		try {
-			JournalTestUtil.addArticle(
-				_group.getGroupId(), 0,
-				ServiceContextTestUtil.getServiceContext(
-					_group.getGroupId(), _user.getUserId(),
-					new long[] {assetCategory.getCategoryId()}));
-
-			Map<String, Object> data = _getData(
-				_getMockLiferayPortletRenderRequest(
-					new String[] {
-						assetVocabulary.getName(),
-						childAssetVocabulary.getName()
-					},
-					LocaleUtil.getSiteDefault()));
-
-			Map<String, Object> props = (Map<String, Object>)data.get("props");
-
-			Assert.assertEquals(
-				JSONUtil.put(
-					JSONUtil.put(
-						"key", String.valueOf(assetCategory.getCategoryId())
-					).put(
-						"name", "category"
-					).put(
-						"value", 1L
-					).put(
-						"vocabularyName", "vocabulary"
-					)
-				).toString(),
-				String.valueOf(props.get("vocabularies")));
-		}
-		finally {
-			_assetVocabularyLocalService.deleteAssetVocabulary(assetVocabulary);
-			_assetVocabularyLocalService.deleteAssetVocabulary(
-				childAssetVocabulary);
-		}
-	}
-
-	@Test
-	public void testGetPropsWithMissingChildAssetVocabularies()
-		throws Exception {
-
-		JournalTestUtil.addArticle(_user.getUserId(), _group.getGroupId(), 0);
-
-		Map<String, Object> data = _getData(
-			_getMockLiferayPortletRenderRequest());
-
-		Map<String, Object> props = (Map<String, Object>)data.get("props");
-
-		Assert.assertEquals(
-			String.valueOf(JSONFactoryUtil.createJSONArray()),
-			String.valueOf(props.get("vocabularies")));
-	}
-
-	@Test
-	public void testGetPropsWithMissingChildAssetVocabulary() throws Exception {
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(
-				_company.getCompanyId(), _company.getGroupId(),
-				_user.getUserId());
-
-		AssetVocabulary assetVocabulary =
-			_assetVocabularyLocalService.fetchGroupVocabulary(
-				serviceContext.getScopeGroupId(), "audience");
-
-		AssetCategory assetCategory = _assetCategoryLocalService.addCategory(
-			_user.getUserId(), _company.getGroupId(),
-			RandomTestUtil.randomString(), assetVocabulary.getVocabularyId(),
-			serviceContext);
-
-		try {
-			JournalTestUtil.addArticle(
-				_group.getGroupId(), 0,
-				ServiceContextTestUtil.getServiceContext(
-					_group.getGroupId(), _user.getUserId(),
-					new long[] {assetCategory.getCategoryId()}));
-
-			Map<String, Object> data = _getData(
-				_getMockLiferayPortletRenderRequest());
-
-			Map<String, Object> props = (Map<String, Object>)data.get("props");
-
-			Assert.assertEquals(
-				JSONUtil.put(
-					JSONUtil.put(
-						"key", String.valueOf(assetCategory.getCategoryId())
-					).put(
-						"name", assetCategory.getTitle(LocaleUtil.US)
-					).put(
-						"value", 1L
-					).put(
-						"vocabularyName",
-						assetVocabulary.getTitle(LocaleUtil.US)
-					)
-				).toString(),
-				String.valueOf(props.get("vocabularies")));
-		}
-		finally {
-			_assetCategoryLocalService.deleteAssetCategory(assetCategory);
-		}
-	}
-
-	@Test
-	public void testGetPropsWithNoneAssetCategory() throws Exception {
-		ServiceContext serviceContext =
-			ServiceContextTestUtil.getServiceContext(
-				_company.getCompanyId(), _company.getGroupId(),
-				_user.getUserId());
-
-		AssetVocabulary assetVocabulary =
-			_assetVocabularyLocalService.fetchGroupVocabulary(
-				serviceContext.getScopeGroupId(), "audience");
-
-		AssetCategory assetCategory = _assetCategoryLocalService.addCategory(
-			_user.getUserId(), _company.getGroupId(),
-			RandomTestUtil.randomString(), assetVocabulary.getVocabularyId(),
-			serviceContext);
-
-		AssetVocabulary childAssetVocabulary =
-			_assetVocabularyLocalService.fetchGroupVocabulary(
-				serviceContext.getScopeGroupId(), "stage");
-
-		AssetCategory childAssetCategory =
-			_assetCategoryLocalService.addCategory(
-				_user.getUserId(), _company.getGroupId(),
-				RandomTestUtil.randomString(),
-				childAssetVocabulary.getVocabularyId(), serviceContext);
-
-		try {
-			JournalTestUtil.addArticle(
-				_group.getGroupId(), 0,
-				ServiceContextTestUtil.getServiceContext(
-					_group.getGroupId(), _user.getUserId(),
-					new long[] {
-						assetCategory.getCategoryId(),
-						childAssetCategory.getCategoryId()
-					}));
-			JournalTestUtil.addArticle(
-				_group.getGroupId(), 0,
-				ServiceContextTestUtil.getServiceContext(
-					_group.getGroupId(), _user.getUserId(),
-					new long[] {childAssetCategory.getCategoryId()}));
-
-			Map<String, Object> data = _getData(
-				_getMockLiferayPortletRenderRequest());
-
-			Map<String, Object> props = (Map<String, Object>)data.get("props");
-
-			Assert.assertEquals(
-				JSONUtil.putAll(
-					JSONUtil.put(
-						"categories",
-						JSONUtil.put(
-							JSONUtil.put(
-								"key",
-								String.valueOf(
-									childAssetCategory.getCategoryId())
-							).put(
-								"name",
-								childAssetCategory.getTitle(LocaleUtil.US)
-							).put(
-								"value", 1L
-							).put(
-								"vocabularyName",
-								childAssetVocabulary.getTitle(LocaleUtil.US)
-							))
-					).put(
-						"key", String.valueOf(assetCategory.getCategoryId())
-					).put(
-						"name", assetCategory.getTitle(LocaleUtil.US)
-					).put(
-						"value", 1L
-					).put(
-						"vocabularyName",
-						assetVocabulary.getTitle(LocaleUtil.US)
-					)
-				).put(
-					JSONUtil.put(
-						"categories",
-						JSONUtil.put(
-							JSONUtil.put(
-								"key",
-								String.valueOf(
-									childAssetCategory.getCategoryId())
-							).put(
-								"name",
-								childAssetCategory.getTitle(LocaleUtil.US)
-							).put(
-								"value", 1L
-							).put(
-								"vocabularyName",
-								childAssetVocabulary.getTitle(LocaleUtil.US)
-							))
-					).put(
-						"key", "none"
-					).put(
-						"name", "No Audience Specified"
-					).put(
-						"value", 1L
-					).put(
-						"vocabularyName",
-						assetVocabulary.getTitle(LocaleUtil.US)
-					)
-				).toString(),
-				String.valueOf(props.get("vocabularies")));
-		}
-		finally {
-			_assetCategoryLocalService.deleteAssetCategory(assetCategory);
-			_assetCategoryLocalService.deleteAssetCategory(childAssetCategory);
-		}
 	}
 
 	@Test
